+++ inkchain-client/lib/Client.js
@@ -86,7 +86,11 @@ var Client = class extends BaseClient {
 
 	/**
 	 * Determine if the inkchain backend is started in
+<<<<<<< HEAD:inkchain-client/lib/Client.js
 	 * [development mode]{@link http://inkchain-inkchain.readthedocs.io/en/latest/Setup/Chaincode-setup.html?highlight=development%20mode}.
+=======
+	 * [development mode]{@link http://hyperledger-inkchain.readthedocs.io/en/latest/Setup/Chaincode-setup.html?highlight=development%20mode}.
+>>>>>>> 2a93d38... 改名inkchain:inkchain-client/lib/Client.js
 	 * In development mode, the endorsing peers will not attempt to spin up a docker instance to run
 	 * the target chaincode requested by a transaction proposal, but instead redirect the invocation
 	 * requests to the chaincode process that has registered itself with the endorsing peer. This makes
@@ -203,7 +207,11 @@ var Client = class extends BaseClient {
 	}
 
 	/**
+<<<<<<< HEAD:inkchain-client/lib/Client.js
 	 * Returns a new {@link TransactionID} object. inkchain transaction ids are constructed
+=======
+	 * Returns a new {@link TransactionID} object. Inkchain transaction ids are constructed
+>>>>>>> 2a93d38... 改名inkchain:inkchain-client/lib/Client.js
 	 * as a hash of a nonce concatenated with the signing identity's serialized bytes. The
 	 * TransactionID object keeps the nonce and the resulting id string bundled together
 	 * as a coherent pair.
@@ -223,7 +231,11 @@ var Client = class extends BaseClient {
 
 	/**
 	 * Extracts the protobuf 'ConfigUpdate' object out of the 'ConfigEnvelope' object
+<<<<<<< HEAD:inkchain-client/lib/Client.js
 	 * that is produced by the [configtxgen tool]{@link http://inkchain-inkchain.readthedocs.io/en/latest/configtxgen.html}.
+=======
+	 * that is produced by the [configtxgen tool]{@link http://hyperledger-inkchain.readthedocs.io/en/latest/configtxgen.html}.
+>>>>>>> 2a93d38... 改名inkchain:inkchain-client/lib/Client.js
 	 * The returned object may then be signed using the signChannelConfig() method of this class. Once the all
 	 * signatures have been collected, the 'ConfigUpdate' object and the signatures may be used
 	 * on the [createChannel()]{@link Client#createChannel} or [updateChannel()]{@link Client#updateChannel} calls.
@@ -310,8 +322,13 @@ var Client = class extends BaseClient {
 	 * @property {byte[]} envelope - Optional. Bytes of the envelope object containing all
 	 *                               required settings and signatures to initialize this channel. This envelope
 	 *                               would have been created by the command line tool
+<<<<<<< HEAD:inkchain-client/lib/Client.js
 	 *                               [configtxgen]{@link http://inkchain-inkchain.readthedocs.io/en/latest/configtxgen.html} or
 	 *                               [configtxlator]{@link https://github.com/inkchain/inkchain/blob/master/examples/configtxupdate/README.md}
+=======
+	 *                               [configtxgen]{@link http://hyperledger-inkchain.readthedocs.io/en/latest/configtxgen.html} or
+	 *                               [configtxlator]{@link https://github.com/hyperledger/inkchain/blob/master/examples/configtxupdate/README.md}
+>>>>>>> 2a93d38... 改名inkchain:inkchain-client/lib/Client.js
 	 * @property {byte[]} config - Optional. Protobuf ConfigUpdate object extracted from a ConfigEnvelope
 	 *                             created by the configtxgen tool. See [extractChannelConfig()]{@link Client#extractChannelConfig}.
 	 *                             The ConfigUpdate object may also be created by the configtxlator tool.
@@ -1023,7 +1040,11 @@ var Client = class extends BaseClient {
 	 * private key and the corresponding x509 certificate. This allows applications
 	 * to use pre-existing crypto materials (private keys and certificates) to
 	 * construct user objects with signing capabilities, as an alternative to
+<<<<<<< HEAD:inkchain-client/lib/Client.js
 	 * dynamically enrolling users with [inkchain-ca]{@link http://inkchain-inkchain-ca.readthedocs.io/en/latest/}
+=======
+	 * dynamically enrolling users with [inkchain-ca]{@link http://hyperledger-inkchain-ca.readthedocs.io/en/latest/}
+>>>>>>> 2a93d38... 改名inkchain:inkchain-client/lib/Client.js
 	 * <br><br>
 	 * Note that upon successful creation of the new user object, it is set to
 	 * the client instance as the current <code>userContext</code>.