+++ test/integration/cloudant-fabricca-tests.js
@@ -24,7 +24,11 @@ var test = _test(tape);
 var path = require('path');
 var Client = require('inkchain-client');
 var User = require('inkchain-client/lib/User.js');
+<<<<<<< HEAD
 var inkchainCAServices = require('inkchain-ca-client/lib/inkchainCAClientImpl');
+=======
+var InkchainCAServices = require('inkchain-ca-client/lib/InkchainCAClientImpl');
+>>>>>>> 2a93d38... 改名inkchain
 var CouchDBKeyValueStore = require('inkchain-client/lib/impl/CouchDBKeyValueStore');
 var testUtil = require('../unit/util.js');
 
@@ -42,10 +46,17 @@ var	tlsOptions = {
 
 // This test first checks to see if a user has already been enrolled. If so,
 // the test terminates. If the user is not yet enrolled, the test uses the
+<<<<<<< HEAD
 // inkchainCAClientImpl to enroll a user, and saves the enrollment materials into the
 // CouchDB KeyValueStore. Then the test uses the Chain class to load the member
 // from the key value store.
 test('Use inkchainCAServices wih a Cloudant CouchDB KeyValueStore', function(t) {
+=======
+// InkchainCAClientImpl to enroll a user, and saves the enrollment materials into the
+// CouchDB KeyValueStore. Then the test uses the Chain class to load the member
+// from the key value store.
+test('Use InkchainCAServices wih a Cloudant CouchDB KeyValueStore', function(t) {
+>>>>>>> 2a93d38... 改名inkchain
 	testUtil.resetDefaults();
 	Client.addConfigFile(path.join(__dirname, 'e2e', 'config.json'));
 	ORGS = Client.getConfigSetting('test-network');
@@ -96,7 +107,11 @@ test('Use inkchainCAServices wih a Cloudant CouchDB KeyValueStore', function(t) 
 					t.end();
 					process.exit(1);
 				}
+<<<<<<< HEAD
 				return new inkchainCAServices(inkchainCAEndpoint, tlsOptions, ORGS[userOrg].ca.name, cryptoSuite);
+=======
+				return new InkchainCAServices(inkchainCAEndpoint, tlsOptions, ORGS[userOrg].ca.name, cryptoSuite);
+>>>>>>> 2a93d38... 改名inkchain
 			},
 			function(err) {
 				t.fail('Error initializing Cloudant KeyValueStore. Exiting.');
@@ -107,7 +122,11 @@ test('Use inkchainCAServices wih a Cloudant CouchDB KeyValueStore', function(t) 
 		.then(
 			function(caService) {
 				logger.debug('ADD: caService - ' + caService);
+<<<<<<< HEAD
 				t.pass('Successfully initialized the inkchain CA service.');
+=======
+				t.pass('Successfully initialized the Inkchain CA service.');
+>>>>>>> 2a93d38... 改名inkchain
 
 				return caService.enroll({
 					enrollmentID: 'admin',
@@ -115,7 +134,11 @@ test('Use inkchainCAServices wih a Cloudant CouchDB KeyValueStore', function(t) 
 				});
 			},
 			function(err) {
+<<<<<<< HEAD
 				t.fail('Failed to initilize the inkchain CA service. Error:');
+=======
+				t.fail('Failed to initilize the Inkchain CA service. Error:');
+>>>>>>> 2a93d38... 改名inkchain
 				logger.error(err.stack ? err.stack : err);
 				t.end();
 			}